# docs/swagger.json
{
  "openapi": "3.1.0",
  "info": {
    "version": "1.0.0",
    "title": "Campers app",
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    },
    "description": "This is a documentation of students app"
  },
  "servers": [
    {
      "url": "https://campersland.netlify.app"
    },
    {
      "url": "http://localhost:3000"
    }
  ],
  "tags": [
    {
      "name": "Campers",
      "description": "Operations with campers."
    }
  ],
  "paths": {
    "/campers": {
      "get": {
        "tags": [
          "Campers"
        ],
        "summary": "Get list of campers",
        "operationId": "getCampers",
        "description": "Get list of campers with this endpoint. Also you can use query parameters for pagination, sort and filtering.",
        "security": [],
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer"
            },
            "example": 1
          },
          {
            "in": "query",
            "name": "perPage",
            "schema": {
              "type": "integer",
              "example": 10
            }
          },
          {
            "in": "query",
            "name": "sortBy",
            "schema": {
              "type": "string",
              "example": "name"
            },
            "description": "All campers' fields can be mentioned. Default value - _id"
          },
          {
            "in": "query",
            "name": "sortOrder",
            "schema": {
              "type": "string",
              "enum": [
                "asc",
                "desc"
              ],
              "example": "asc"
            }
          },
          {
            "in": "query",
            "name": "location",
            "schema": {
              "type": "string",
              "example": "Ukraine, Kyiv"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully found campers!",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "status",
                    "message",
                    "data"
                  ],
                  "properties": {
                    "status": {
                      "type": "integer",
                      "example": 200
                    },
                    "message": {
                      "type": "string",
                      "example": "Successfully found campers!"
                    },
                    "data": {
                      "type": "object",
                      "required": [
                        "data",
                        "page",
                        "perPage",
                        "totalItems",
                        "totalPages",
                        "hasPreviousPage",
                        "hasNextPage"
                      ],
                      "properties": {
                        "data": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/camper"
                          }
                        },
                        "page": {
                          "type": "number",
                          "example": 2
                        },
                        "perPage": {
                          "type": "number",
                          "example": 4
                        },
                        "totalItems": {
                          "type": "number",
                          "example": 4
                        },
                        "totalPages": {
                          "type": "number",
                          "example": 4
                        },
                        "hasPreviousPage": {
                          "type": "boolean",
                          "example": true
                        },
                        "hasNextPage": {
                          "type": "boolean",
                          "example": false
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/404"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "details": {
        "type": "object",
        "required": [
          "airConditioner",
          "bathroom",
          "kitchen",
          "beds",
          "TV",
          "CD",
          "radio",
          "shower",
          "toilet",
          "freezer",
          "hob",
          "microwave",
          "gas",
          "water"
        ],
        "properties": {
          "airConditioner": {
            "type": "integer",
            "example": 1
          },
          "bathroom": {
            "type": "integer",
            "example": 1
          },
          "kitchen": {
            "type": "integer",
            "example": 1
          },
          "beds": {
            "type": "integer",
            "example": 3
          },
          "TV": {
            "type": "integer",
            "example": 1
          },
          "CD": {
            "type": "integer",
            "example": 1
          },
          "radio": {
            "type": "integer",
            "example": 1
          },
          "shower": {
            "type": "integer",
            "example": 1
          },
          "toilet": {
            "type": "integer",
            "example": 1
          },
          "freezer": {
            "type": "integer",
            "example": 1
          },
          "hob": {
            "type": "integer",
            "example": 3
          },
          "microwave": {
            "type": "integer",
            "example": 1
          },
          "gas": {
            "type": "string",
            "example": "35kg"
          },
          "water": {
            "type": "string",
            "example": "151l"
          }
        }
      },
      "review": {
        "type": "object",
        "required": [
          "reviewer_name",
          "reviewer_rating",
          "comment"
        ],
        "properties": {
          "reviewer_name": {
            "type": "string",
            "example": "Alice"
          },
          "reviewer_rating": {
            "type": "integer",
            "minimum": 0,
            "maximum": 5,
            "example": 5
          },
          "comment": {
            "type": "string",
            "example": "Exceptional RV! Highly recommended!"
          }
        }
      },
      "camper": {
        "type": "object",
        "required": [
          "name",
          "price",
          "rating",
          "location",
          "adults",
          "children",
          "engine",
          "transmission",
          "form",
          "length",
          "width",
          "height",
          "tank",
          "consumption",
          "description",
          "details",
          "gallery",
          "reviews"
        ],
        "properties": {
          "_id": {
            "description": "Contact's id",
            "type": "string",
            "example": "665c7107b4570a0fe55ea884"
          },
          "name": {
            "description": "Contact's name",
            "type": "string",
            "example": "Yulia Shevchenko"
          },
          "price": {
            "description": "Camper's price",
            "type": "number",
            "example": "10000"
          },
          "rating": {
            "description": "Rating of camper",
            "type": "number",
            "minimum": 0,
            "maximum": 5,
            "example": 4
          },
          "location": {
            "description": "Location of camper",
            "type": "string",
            "example": "Ukraine, Kyiv"
          },
          "adults": {
            "description": "Number of seats for adults",
            "type": "integer",
            "example": 2
          },
          "children": {
            "description": "Number of seats for children",
            "type": "integer",
            "example": 1
          },
          "engine": {
            "description": "Type of engine's fuels",
            "type": "string",
            "enum": [
              "petrol",
              "diesel",
              "hybrid",
              "electricity"
            ],
            "example": "petrol"
          },
          "transmission": {
            "description": "Type of transmission",
            "type": "string",
            "enum": [
              "automatic",
              "mechanic"
            ],
            "example": "automatic"
          },
          "form": {
            "description": "Form of camper",
            "type": "string",
            "enum": [
              "alcove",
              "fullyIntegrated",
              "panelTruck"
            ],
            "example": "alcove"
          },
          "length": {
            "description": "Length of camper",
            "type": "string",
            "example": "5.4m"
          },
          "width": {
            "description": "Width of camper",
            "type": "string",
            "example": "2.01m"
          },
          "height": {
            "description": "Height of camper",
            "type": "string",
            "example": "2.05m"
          },
          "tank": {
            "description": "Amount of fuel in liters",
            "type": "string",
            "example": "208l"
          },
          "consumption": {
            "description": "Amount of liters per kilometer",
            "type": "string",
            "example": "30l/100km"
          },
          "description": {
            "description": "Descriptionf of camper",
            "type": "string",
            "example": "Embark on an unforgettable journey with the Road Bear C 23-25, an epitome of comfort and convenience on wheels."
          },
          "details": {
            "$ref": "#/components/schemas/details"
          },
          "gallery": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "example": [
              "https://ftp.goit.study/img/campers-test-task/1-1.webp",
              "https://ftp.goit.study/img/campers-test-task/1-2.webp",
              "https://ftp.goit.study/img/campers-test-task/1-3.webp"
            ]
          },
          "reviews": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/review"
            }
          }
        }
      }
    },
    "responses": {
      "404": {
        "description": "Campers not found",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "required": [
                "status",
                "message",
                "data"
              ],
              "properties": {
                "status": {
                  "type": "integer",
                  "example": 404
                },
                "message": {
                  "type": "string",
                  "example": "Campers not found"
                },
                "data": {
                  "type": "object",
                  "required": [
                    "message"
                  ],
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "ContCampersact not found"
                    }
                  }
                }
              }
            }
          }
        }
      }
    }
  }
}